Added solution to problem 36 and moved string functions to math header.

+++ src/problems/1-100/1-20/p1.cpp
@@ -2,8 +2,8 @@
 
 int64_t problem_1()
 {
-    int sum = 0;
-    for(int i = 1; i < 1000; ++i) {
+    int64_t sum = 0;
+    for(int64_t i = 1; i < 1000; ++i) {
         if(i % 3 == 0 || i % 5 == 0) {
             sum += i;
         }
